ChatApp/requirements.txt
asgiref==3.2.3
certifi==2019.11.28 
chardet==3.0.4 
CodeConvert==3.0.2 
Django==2.2.6 
django-admin==2.0.0 
django-excel-response2==2.0.8 
django-models-ext==1.1.9 
django-short-url==1.1.4 
django-six==1.0.4 
django-widget-tweaks==1.4.1 
djangorestframework==3.9.1 
djangorestframework-jwt==1.11.0
furl==2.1.0 
idna==2.8
orderedmultidict==1.0.1 
psycopg2==2.8.4 
PyJWT==1.7.1 
python-dateutil==2.8.1 
python-memcached==1.59
pytz==2017.3 
requests==2.22.0 
screen==1.0.1 
shortuuid==0.5.0 
six==1.13.0 
sqlparse==0.3.0 
TimeConvert==1.5.1
tzlocal==2.0.0
urllib3==1.25.7 
validate-email==1.3 
xlwt==1.3.0
